requirements.txt
academictorrents==2.3.3
accelerate==0.28.0
albumentations==1.3.1
augly==1.0.0
CairoSVG==2.7.1
diffusers==0.26.3
einops==0.7.0
fire==0.6.0
fsspec==2023.4.0
imwatermark==0.0.2
intel_extension_for_pytorch==2.2.0
kornia==0.7.0
lightning==2.1.0
matplotlib==3.8.0
more_itertools==10.2.0
natsort==8.4.0
numpy==1.24.1
omegaconf==2.3.0
open_clip_torch==2.22.0
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
packaging==24.0
pandas==2.2.1
Pillow==9.3.0
Pillow==10.0.1
Pillow==10.2.0
pipelines==0.0.14
pudb==2024.1
pytest==8.1.1
pytorch_fid==0.3.0
pytorch_lightning==2.1.0
PyYAML==6.0.1
PyYAML==6.0.1
pyzbar==0.1.9
qrcode==7.4.2
Requests==2.31.0
safetensors==0.4.2
scikit_learn==1.3.1
scipy==1.12.0
setuptools==68.0.0
skimage==0.0
streamlit_drawable_canvas==0.9.3
timm==0.9.8
torch==2.1.0+cu121
torchdata==0.7.1
torchvision==0.16.0+cu121
tqdm==4.66.1
transformers==4.34.1
typing_extensions==4.10.0
webdataset==0.2.86
xformers==0.0.24
